requirements.txt
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
asttokens==2.0.8
attrs==22.1.0
av==9.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
docutils==0.15.2
entrypoints==0.4
executing==0.10.0
gitdb==4.0.9
GitPython==3.1.27
google-crc32c==1.3.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.12.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
netifaces==0.11.0
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pycparser==2.21
pydeck==0.7.1
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
requests==2.28.1
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.4.0
streamlit==1.11.0
streamlit-webrtc==0.43.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
widgetsnbextension==4.0.1
zipp==3.8.1
